+++ app/build.gradle
@@ -78,12 +78,6 @@ android {
 
         manifestPlaceholders += ['appAuthRedirectScheme': 'me.banes.chris.tivi']
 
-        javaCompileOptions {
-            annotationProcessorOptions {
-                arguments = ["room.schemaLocation": "$projectDir/schemas".toString()]
-            }
-        }
-
         playAccountConfig = playAccountConfigs.defaultAccountConfig
 
         // Crashlytics. Disabled by default
@@ -135,7 +129,7 @@ android {
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(":data")
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jre8:${versions.kotlin}"
 
@@ -143,10 +137,6 @@ dependencies {
     implementation "android.arch.lifecycle:extensions:1.0.0-rc1"
     implementation "android.arch.lifecycle:reactivestreams:1.0.0-rc1"
     kapt "android.arch.lifecycle:compiler:1.0.0-rc1"
-
-    implementation "android.arch.persistence.room:runtime:1.0.0-rc1"
-    implementation "android.arch.persistence.room:rxjava2:1.0.0-rc1"
-    kapt "android.arch.persistence.room:compiler:1.0.0-rc1"
 
     implementation "android.arch.paging:runtime:1.0.0-alpha3"
 
@@ -156,7 +146,6 @@ dependencies {
 
     implementation "com.android.support.constraint:constraint-layout:${versions.constraint_layout}"
 
-    implementation 'org.threeten:threetenbp:1.3.6:no-tzdb'
     implementation 'com.gabrielittner.threetenbp:lazythreetenbp:0.2.0'
 
     implementation("com.uwetrottmann.trakt5:trakt-java:5.7.3", {
@@ -171,7 +160,6 @@ dependencies {
     implementation "io.reactivex.rxjava2:rxjava:${versions.rx}"
     implementation "io.reactivex.rxjava2:rxandroid:2.0.1"
 
-    implementation "com.google.dagger:dagger:${versions.dagger}"
     implementation "com.google.dagger:dagger-android:${versions.dagger}"
     implementation "com.google.dagger:dagger-android-support:${versions.dagger}"
     kapt "com.google.dagger:dagger-compiler:${versions.dagger}"
@@ -200,11 +188,6 @@ dependencies {
     implementation('com.crashlytics.sdk.android:crashlytics:2.7.1@aar') {
         transitive = true
     }
-}
-
-String propOrDef(String propertyName, String defaultValue) {
-    def propertyValue = project.properties[propertyName]
-    return propertyValue != null ? propertyValue : defaultValue
 }
 
 def getGitHash() {
