+++ app/src/main/java/me/banes/chris/tivi/calls/PopularCall.kt
@@ -22,6 +22,7 @@ import com.uwetrottmann.trakt5.enums.Extended
 import io.reactivex.Maybe
 import io.reactivex.Single
 import me.banes.chris.tivi.api.ItemWithIndex
+import me.banes.chris.tivi.data.DatabaseTxRunner
 import me.banes.chris.tivi.data.daos.PopularDao
 import me.banes.chris.tivi.data.daos.TiviShowDao
 import me.banes.chris.tivi.data.entities.PopularEntry
@@ -29,7 +30,6 @@ import me.banes.chris.tivi.data.entities.PopularListItem
 import me.banes.chris.tivi.data.entities.TiviShow
 import me.banes.chris.tivi.extensions.toRxSingle
 import me.banes.chris.tivi.util.AppRxSchedulers
-import me.banes.chris.tivi.util.DatabaseTxRunner
 import javax.inject.Inject
 
 class PopularCall @Inject constructor(
