+++ app/src/main/java/me/banes/chris/tivi/inject/AppComponent.kt
@@ -20,6 +20,7 @@ import dagger.Component
 import dagger.android.AndroidInjector
 import dagger.android.support.AndroidSupportInjectionModule
 import me.banes.chris.tivi.TiviApplication
+import me.banes.chris.tivi.data.DatabaseModule
 import me.banes.chris.tivi.home.HomeBuilder
 import me.banes.chris.tivi.settings.SettingsBuilder
 import me.banes.chris.tivi.trakt.TraktModule
