+++ build.gradle
@@ -66,3 +66,8 @@ allprojects {
 task clean(type: Delete) {
     delete rootProject.buildDir
 }
+
+String propOrDef(String propertyName, String defaultValue) {
+    def propertyValue = project.properties[propertyName]
+    return propertyValue != null ? propertyValue : defaultValue
+}
