+++ data/build.gradle
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2017 Google, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+apply plugin: 'com.android.library'
+
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-kapt'
+
+android {
+    compileSdkVersion 27
+
+    defaultConfig {
+        minSdkVersion 23
+
+        javaCompileOptions {
+            annotationProcessorOptions {
+                arguments = ["room.schemaLocation": "$projectDir/schemas".toString()]
+            }
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+}
+
+dependencies {
+    api project(":base")
+
+    implementation "android.arch.persistence.room:runtime:1.0.0-rc1"
+    implementation "android.arch.persistence.room:rxjava2:1.0.0-rc1"
+    kapt "android.arch.persistence.room:compiler:1.0.0-rc1"
+
+    kapt "com.google.dagger:dagger-compiler:${versions.dagger}"
+
+    api "android.arch.paging:runtime:1.0.0-alpha3"
+
+    api 'org.threeten:threetenbp:1.3.6:no-tzdb'
+}
