+++ data/src/main/java/me/banes/chris/tivi/data/Entry.kt
