+++ data/src/main/java/me/banes/chris/tivi/data/entities/TrendingEntry.kt
